# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Discography Management API
  description: API for managing artists and albums in a discography system
  version: 1.0.0
  contact:
    name: Marcos Oliveira
    email: marcos@example.com
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/artists/album/{id}:
    get:
      tags:
        - Artists
      summary: Find artists by album ID
      description: Find all artists associated with a specific album
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: Album UUID
            example: "550e8400-e29b-41d4-a716-446655440002"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of artists associated with the album
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArtistResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
...
  /api/v1/albums:
    get:
      tags:
        - Albums
      summary: Get paginated list of albums
      description: Retrieve a paginated list of all albums
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of albums
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlbumResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean

    post:
      tags:
        - Albums
      summary: Create new album
      description: Create a new album with optional cover image upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - request
              properties:
                request:
                  type: object
                  description: Album request data
                  required:
                    - title
                    - releaseDate
                    - artistIds
                  properties:
                    title:
                      type: string
                      description: Album title
                      example: "Abbey Road"
                    releaseDate:
                      type: string
                      format: date
                      description: Album release date
                      example: "1969-09-26"
                    artistIds:
                      type: array
                      items:
                        type: string
                        format: uuid
                      description: List of artist UUIDs
                      example: ["550e8400-e29b-41d4-a716-446655440000", "550e8400-e29b-41d4-a716-446655440001"]
                file:
                  type: string
                  format: binary
                  description: Album cover image (optional)
                  example: "cover.jpg"
      responses:
        '201':
          description: Album created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponseDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
  /api/v1/albums/{id}:
    get:
      tags:
        - Albums
      summary: Get album by ID
      description: Retrieve a specific album by its UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
      responses:
        '200':
          description: Album details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponseDTO'
        '404':
          description: Album not found

    put:
      tags:
        - Albums
      summary: Update album
      description: Update an existing album's information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumRequestDTO'
      responses:
        '204':
          description: Album updated successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'

    delete:
      tags:
        - Albums
      summary: Delete album
      description: Delete an album by its UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
      responses:
        '204':
          description: Album deleted successfully
  /api/v1/albums/title/{title}:
    get:
      tags:
        - Albums
      summary: Find albums by title
      description: Search for albums by title with pagination support
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
            description: Album title to search for
            example: "Abbey Road"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of matching albums
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlbumResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
  /api/v1/albums/releaseDate/:
    get:
      tags:
        - Albums
      summary: Find albums by release date range
      description: Search for albums by release date range with pagination support
      parameters:
        - name: after
          in: query
          required: false
          schema:
            type: string
            format: date
            description: Filter for albums released after this date
            example: "1969-01-01"
        - name: before
          in: query
          required: false
          schema:
            type: string
            format: date
            description: Filter for albums released before this date
            example: "1970-12-31"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of albums within the specified date range
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlbumResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
  /api/v1/albums/artist/{id}:
    get:
      tags:
        - Albums
      summary: Find albums by artist ID
      description: Find all albums associated with a specific artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: Artist UUID
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of albums by the specified artist
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlbumResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
  /api/v1/albums/{id}/cover:
    put:
      tags:
        - Albums
      summary: Upload album cover
      description: Upload or update a cover image for an album
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440002"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Cover image file
                  example: "cover.jpg"
      responses:
        '204':
          description: Cover image uploaded successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
components:
  schemas:
...
    ArtistResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "The Beatles"
        genre:
          type: string
          example: "Rock"
        profileImageUrl:
          type: string
          example: "http://localhost:8080/api/v1/artists/550e8400-e29b-41d4-a716-446655440000/profile-image"
        albumTitles:
          type: array
          items:
            type: string
          example: ["Abbey Road", "Let It Be", "Revolver"]
          minItems: 0
          maxItems: 100
    AlbumRequestDTO:
      type: object
      required:
        - title
        - releaseDate
        - artistIds
      properties:
        title:
          type: string
          description: Album title
          example: "Abbey Road"
          minLength: 1
          pattern: "^[a-zA-Z0-9\\s\\-\\'\\.]+$"
        releaseDate:
          type: string
          format: date
          description: Album release date
          example: "1969-09-26"
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        artistIds:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
            format: uuid
          description: List of artist UUIDs
          example: ["550e8400-e29b-41d4-a716-446655440000", "550e8400-e29b-41d4-a716-446655440001"]
    AlbumResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
        title:
          type: string
          example: "Abbey Road"
        releaseDate:
          type: string
          format: date
          example: "1969-09-26"
        coverUrl:
          type: string
          example: "http://localhost:8080/api/v1/albums/550e8400-e29b-41d4-a716-446655440002/cover"
        artistNames:
          type: array
          items:
            type: string
          example: ["The Beatles"]
          minItems: 1
          maxItems: 10
    ErrorDetails:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2024-01-26T10:30:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Artist not found with id: 550e8400-e29b-41d4-a716-446655440000"
        path:
          type: string
          example: "/api/v1/artists/550e8400-e29b-41d4-a716-446655440000"
      required:
        - timestamp
        - status
        - error
        - message
        - path
...